requirements.txt
lxml==3.8.0
ftfy==5.0.2
autobahn==17.9.3
beautifulsoup4==4.6.0
PyYAML==3.12
